@@ -5,12 +5,12 @@
   "keywords": [],
   "main": "src/index.js",
   "dependencies": {
-    "react": "16.8.6",
-    "react-dom": "16.8.6",
-    "react-scripts": "3.2.0"
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-scripts": "5.0.1"
   },
   "devDependencies": {
-    "typescript": "3.3.3"
+    "typescript": "4.8.3"
   },
   "scripts": {
     "start": "react-scripts start",
